added functions perms decor

+++ iam/iam.py
@@ -12,6 +12,8 @@ iam = Blueprint('iam', __name__)
 @auth.login_required(perms=[11])
 def creatRole():
     data = request.get_json()
+    if 'roleName' not in data or 'permissions' not in data:
+        abort(400, "Missing data")
     db.createRole(data)
     return {}
 
@@ -20,5 +22,7 @@ def creatRole():
 @auth.login_required(perms=[11])
 def removeRole():
     data = request.get_json()
+    if 'roleId' not in data:
+        abort(400, "Missing data")
     db.removeRole(data['roleId'])
     return {}
